requirements.txt
asgiref==3.8.1
Django==5.1.2
sqlparse==0.5.1
typing_extensions==4.12.2
openai==0.28
djangorestframework==3.15.2
psycopg2-binary==2.9.9
requests==2.32.3
django-cors-headers==4.4.0
assemblyai==0.34.0
pyannote.audio==3.3.2
pyannote.audio==3.3.1
openai==1.51.2
regex==2024.9.11
ffmpeg-python==0.2.0
pydub==0.25.1
fpdf==1.7.2